Refactored tiler class, added support for EPT tiling
